+++ resources/views/livewire/relationship-schema/belongsto.blade.php
@@ -2,7 +2,7 @@
 
 <div class="form-group col-md-3 mb-1 px-1">
     <label class="mb-1">Column</label>
-    @if($relationship['created_by_column'])
+    @if(isset($relationship['created_by_column']) && $relationship['created_by_column'])
     <input
         type="text"
         name="relationships[{{ $relationship_index }}][relationship_column]"